Directly return the id if passed as values
In DatabaseModel, on creation, we try to return the autoincremented id.
Problem is primary key is not always an autoincremented value. In this
situation, the value should passed as an argument and can be directly
returned instead of calling the `lastInsertId` method.

Otherwise, with SQLite at least, the value wasn't consistent.