imp: Return pointer in Url::for if undeclared
We might want to return a fake URL, or the path to a file in the
public/ folder. It might hide a broken URL and it is indeed a risk. In
the mean time, it means a site might look less broken in production (but
it is a very good reason).

If it becomes a problem, I'll look for another solution.