Change the ActionController system
The controllers actions were previously declared as functions within a
complex namespace. It made the whole system a bit weird for a PHP
project.

Actions are now declared in a controller, which is declared as a class.
By default, controllers are at the root of the `src/` folder, but can be
placed deeper thanks to commit c546235.