fix: Handle router patterns longer than paths
This bug was introduced in b69a786. Before, we were checking both
strings were the same length before the for-loop, so they had the same
number of elements. With the introduction of the wildcard parameter, we
must iterate immediately since the wildcard can match several
elements of the path. It means we also need to check the path element
exists before using it!