Fix how URI are handled in Request
Errors were raised when people tried to access URLs such as
`http://domain.com//test` (note the two slashes). Indeed, `REQUEST_URI`
was set to `//test`, which is badly interpreted by the `parse_url()`
function to be the domain name. Then, the `$uri_components['path']`
variable was missing and the request failed with a "cannot be parsed"
error.