Rework the Database class
PDO connections are closed by unassigning their references. The Database
instance was inheriting PDO class. It was also passed to several
DatabaseModel, and so was referenced multiple times. The consequence was
it was difficult to close the PDO connection.

Now, Database is a wrapper around a PDO object. It also provides a
`close()` method which simply unassign the unique variable referencing
the PDO object.

It should solve the issue where tests were failing in chain with a
`database "xxx" is being accessed by other users` error.

Reference: https://www.php.net/manual/pdo.connections.php