tec: Change ViewHelpers functions to static methods
This allows these methods to be used outside of view files. In
particular, it allows us to test the methods.

The function are still available of course, but they only call the
methods.

Breaking change: the `formatDate` method (i.e. `_d`) now uses the
`IntlDateFormatter` class to translate and format dates. Thus, the
format syntax has changed.

Also, the `csrf_token` function is now deprecated and CSRF token should
be passed as a variable to the views.