tec: Allow models to be inherited
Recordable wasn't using the correct class name for inherited models. It
was a problem for Jobs because Recordable was loading the \Minz\Job class
(which is abstract) instead of the final class.

We still use the initial model (i.e. `self::class`) when we are looking
for the Table attribute. Otherwise, PHP isn't able to find it.